+++ addons/website_sale_stock/tests/test_website_sale_stock_meta.py
@@ -33,7 +33,7 @@ class TestWebsiteSaleStockMeta(WebsiteSaleFeedCommon):
         self.assertEqual('out of stock', self.items[self.blanket]['availability'])
         self.assertEqual('in stock', self.red_sofa_item['availability'])  # allow out_of_stock
 
-    def test_meta_stock_info_in_stock_with_group(self):
+    def test_meta_items_keep_website_stock_separate(self):
         self.blue_sofa.allow_out_of_stock_order = False
         # setup second website with seperate stock
         warehouse_2 = self.env['stock.warehouse'].create({'name': 'Stock 2', 'code': 'WH2'})
